Move signaling to a separate file
